+++ src/model.tsx
@@ -1,8 +1,11 @@
-import { AlignHorizontalLeftIcon, MaleIcon } from './Icons'
 import { types, Instance } from 'mobx-state-tree'
+import VisibilityIcon from '@material-ui/icons/Visibility'
 import { ElementId } from '@jbrowse/core/util/types/mst'
 import { MenuItem } from '@jbrowse/core/ui'
 import PluginManager from '@jbrowse/core/PluginManager'
+import { AlignHorizontalLeftIcon, HourglassIcon, MaleIcon } from './Icons'
+import FolderOpenIcon from '@material-ui/icons/FolderOpen'
+import { FileLocation } from '@jbrowse/core/util/types'
 
 export default function IdeogramView(pluginManager: PluginManager) {
   return types
@@ -10,10 +13,23 @@ export default function IdeogramView(pluginManager: PluginManager) {
       type: types.literal('IdeogramView'),
       displayName: types.maybe(types.string),
       id: ElementId,
+
+      // ideogram config
       sex: 'female',
       orientation: 'vertical',
-      pliody: 2,
+      region: '1',
+      assembly: 'hg38',
+
+      // display options
+      allRegions: false,
+      showImportForm: true,
+      showAnnotations: true,
     })
+    .volatile(() => ({
+      annotationsLocation: (undefined as unknown) as FileLocation,
+      widgetAnnotations: (undefined as unknown) as object,
+      ideoAnnotations: (undefined as unknown) as object,
+    }))
     .actions(self => ({
       setWidth(n: number) {
         /* do nothing */
@@ -21,13 +37,46 @@ export default function IdeogramView(pluginManager: PluginManager) {
       setDisplayName(str: string) {
         self.displayName = str
       },
+      setRegion(chr: string) {
+        self.region = chr.split('chr')[1]
+      },
+      setAssembly(asm: string) {
+        self.assembly = asm
+      },
+      setAllRegions(toggle: boolean) {
+        self.allRegions = toggle
+      },
+      setOrientation(ori: string) {
+        if (ori === 'horizontal') {
+          self.orientation = ori
+        }
+        if (ori === 'vertical') {
+          self.orientation = ori
+        }
+      },
+      setShowImportForm(toggle: boolean) {
+        self.showImportForm = toggle
+      },
+      setAnnotationsLocation(loc: FileLocation) {
+        self.annotationsLocation = loc
+      },
+      setWidgetAnnotations(obj: object) {
+        self.widgetAnnotations = obj
+      },
+      setIdeoAnnotations(obj: object) {
+        self.ideoAnnotations = obj
+      },
+      toggleAllRegions(toggle: boolean) {
+        if (toggle === false) {
+          this.setOrientation('horizontal')
+        }
+        this.setAllRegions(toggle)
+      },
       toggleOrientation() {
         if (self.orientation === 'horizontal') {
-          self.orientation = 'vertical'
-          self.pliody = 2
+          this.setOrientation('vertical')
         } else {
-          self.orientation = 'horizontal'
-          self.pliody = 1
+          this.setOrientation('horizontal')
         }
       },
       toggleSex() {
@@ -37,14 +86,30 @@ export default function IdeogramView(pluginManager: PluginManager) {
           self.sex = 'male'
         }
       },
+      toggleAnnotations() {
+        self.showAnnotations = !self.showAnnotations
+      },
     }))
     .views(self => ({
       menuItems(): MenuItem[] {
         const menuItems: MenuItem[] = [
           {
+            label: 'Return to import form',
+            icon: FolderOpenIcon,
+            onClick: () => self.setShowImportForm(true),
+          },
+          {
+            label: 'Show all regions in assembly',
+            icon: VisibilityIcon,
+            type: 'checkbox',
+            checked: self.allRegions === true,
+            onClick: () => self.toggleAllRegions(!self.allRegions),
+          },
+          {
             label: 'Horizontal Display',
             icon: AlignHorizontalLeftIcon,
             type: 'checkbox',
+            disabled: self.allRegions === false,
             checked: self.orientation === 'horizontal',
             onClick: () => self.toggleOrientation(),
           },
@@ -54,6 +119,14 @@ export default function IdeogramView(pluginManager: PluginManager) {
             type: 'checkbox',
             checked: self.sex === 'male',
             onClick: () => self.toggleSex(),
+          },
+          {
+            label: 'Show annotations',
+            icon: HourglassIcon,
+            type: 'checkbox',
+            checked: self.showAnnotations === true,
+            disabled: self.widgetAnnotations === undefined,
+            onClick: () => self.toggleAnnotations(),
           },
         ]
         return menuItems