+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 - Moved File > Add > Ideogram View to Add > Ideogram View
 - Applied default view title to view
 - Added View Menu toggles for sex and (figure) orientation
+- Added configuration menu before opening the view for assembly and region
+- Added ability to view single region of ideogram as well as all regions
+- Added View Menu toggle for 'Show all Regions' and 'Return to import'
+- Added FileSelector to accept annotation files
+- Added an 'adapter' to process TSV files for annotations (note: to annotate this file needs to have certain features, see the README for details)
+- Added IdeogramFeatureWidget to display annotation data that has been clicked
+- Added ability to view or hide annotations that have been uploaded to the ideogram
 
 # v1.0.0